requirements-gpu.txt
-r ./requirements.txt

tensorflow-gpu==2.0.0a0